requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
plotly>=5.17.0
pyarrow>=12.0.0

